Fix NormalizePath trailing slash appending
NormalizePath would append a trailing slash if and only if it would
merge multiple slashes into one. For example, the following path would
not get a trailing slash appended:

"/sample/path"

On the other hand, this path would get a trailing slash appended:

"/sample//path"

After this commit, the trailing slash is appended for both the paths.